+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ## v1.0.0-beta.1
 
 - Added `CollectionDiff::setNormalizers`
+- Added CircleCI
+- Updated `CollectionDiff::_construct` first param is now an array of `NormalizerInterface` 
 - Updated `README.md`